Add message that products did not found on Product List page when there are no results and fix bug with search bar in AppBar component
Now, when user trying to find products with wrong filters user will get message that products were not found.
And also was a bug in search bar in AppBar component.
When user erase all his input in search it was trying to find the products with the query that contains last value before input erasing.